+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - **New data structure** resilient to information loss because of failed writes on some cards
+- Use external EEPROM AT24C32 onboard DS3231 RTC
 - Fix latency calculation in sync_clock.py, wait for the prompt explicitly
 - Fix layout in Android app for smaller screens
 - Keep current view when the screen is rotated